Moves functions to components/core/ms-utils.ts
