+++ client/src/components/core/ms-modal/MsPasswordInputModal.vue
@@ -25,7 +25,7 @@ import { ref } from 'vue';
 import MsModal from '@/components/core/ms-modal/MsModal.vue';
 import MsPasswordInput from '@/components/core/ms-input/MsPasswordInput.vue';
 import { MsModalResult } from '@/components/core/ms-types';
-import { GetPasswordOptions } from '@/common/inputs';
+import { GetPasswordOptions } from '@/components/core/ms-utils';
 
 defineProps<GetPasswordOptions>();
